Refactoring to clean up method naming
