+++ src/handlers/RegisterNewAnimalHandler.cs
@@ -20,16 +20,12 @@ namespace PetHome.Handlers
 
         public Task<Unit> Handle(RegisterNewAnimal request, CancellationToken cancellationToken)
         {
-            if (request == null || string.IsNullOrEmpty(request.Species) ||
-                string.IsNullOrEmpty(request.Breed) ||
-                string.IsNullOrEmpty(request.Color) ||
-                string.IsNullOrEmpty(request.Gender) ||
-                request.Weight == 0)
+            if (DoesNotHaveMinimumRequiredAnimalInformation(request))
             {
                 throw new ArgumentException();
             }
 
-            _context.Add(new Animal
+            _context.Insert(new Animal
             {
                 Species = request.Species,
                 Color = request.Color,
@@ -44,6 +40,15 @@ namespace PetHome.Handlers
             });
 
             return Unit.Task;
+        }
+
+        private bool DoesNotHaveMinimumRequiredAnimalInformation(RegisterNewAnimal request)
+        {
+            return request == null || string.IsNullOrEmpty(request.Species) ||
+                   string.IsNullOrEmpty(request.Breed) ||
+                   string.IsNullOrEmpty(request.Color) ||
+                   string.IsNullOrEmpty(request.Gender) ||
+                   request.Weight == 0;
         }
     }
 }
